# openapi.yaml
openapi: 3.1.0
info:
  title: PeoplePay360 API
  version: 1.0.0
  description: >
    Integrated HR and payroll API. Employees, contracts, working schedules, attendance,
    time off, salary structures and rules, pay runs, payslips, dashboard and reports.
    All money values are decimal strings. All calendar dates are YYYY-MM-DD strings.
servers:
  - url: http://localhost:4000
    description: Local
...
security:
  - sessionCookie: []
paths:
...
  /api/time-off/types:
    get:
      tags: [timeoff]
      operationId: listTimeOffTypes
      summary: Time off type list
      parameters:
        - { $ref: '#/components/parameters/ActiveOnly' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Type list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffTypeListResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [timeoff]
      operationId: createTimeOffType
      summary: Create a time off type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTimeOffTypeRequest' }
      responses:
        '201':
          description: Type created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffTypeResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/types/{id}:
    get:
      tags: [timeoff]
      operationId: getTimeOffType
      summary: Single time off type
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Type detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffTypeResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [timeoff]
      operationId: updateTimeOffType
      summary: Update a time off type
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTimeOffTypeRequest' }
      responses:
        '200':
          description: Type updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffTypeResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/time-off/allocations:
    get:
      tags: [timeoff]
      operationId: listAllocations
      summary: Allocations with derived taken and remaining
      parameters:
        - { $ref: '#/components/parameters/EmployeeId' }
        - { $ref: '#/components/parameters/TimeOffTypeId' }
        - { $ref: '#/components/parameters/AllocationStatusFilter' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Allocation list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationListResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [timeoff]
      operationId: createAllocation
      summary: Create an allocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAllocationRequest' }
      responses:
        '201':
          description: Allocation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/time-off/allocations/{id}:
    get:
      tags: [timeoff]
      operationId: getAllocation
      summary: Single allocation
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Allocation detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [timeoff]
      operationId: updateAllocation
      summary: Update a draft allocation
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAllocationRequest' }
      responses:
        '200':
          description: Allocation updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/allocations/{id}/approve:
    post:
      tags: [timeoff]
      operationId: approveAllocation
      summary: Approve an allocation, making the balance available
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Allocation approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/allocations/{id}/refuse:
    post:
      tags: [timeoff]
      operationId: refuseAllocation
      summary: Refuse an allocation, withdrawing the balance
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Allocation refused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/requests:
    get:
      tags: [timeoff]
      operationId: listTimeOffRequests
      summary: Paginated, filterable request list
      parameters:
        - { $ref: '#/components/parameters/EmployeeId' }
        - { $ref: '#/components/parameters/TimeOffTypeId' }
        - { $ref: '#/components/parameters/RequestStatusFilter' }
        - { $ref: '#/components/parameters/DateFrom' }
        - { $ref: '#/components/parameters/DateTo' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Request list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestListResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [timeoff]
      operationId: createTimeOffRequest
      summary: Raise a time off request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTimeOffRequestRequest' }
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/requests/{id}:
    get:
      tags: [timeoff]
      operationId: getTimeOffRequest
      summary: Single request with its consumed allocation
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Request detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [timeoff]
      operationId: updateTimeOffRequest
      summary: Edit or cancel a request awaiting approval
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTimeOffRequestRequest' }
      responses:
        '200':
          description: Request updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/requests/{id}/approve:
    post:
      tags: [timeoff]
      operationId: approveTimeOffRequest
      summary: Approve a request and consume allocation balance
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Request approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestResponse' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/requests/{id}/refuse:
    post:
      tags: [timeoff]
      operationId: refuseTimeOffRequest
      summary: Refuse a request and restore any consumed balance
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefuseTimeOffRequestRequest' }
      responses:
        '200':
          description: Request refused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffRequestResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/time-off/dashboard:
    get:
      tags: [timeoff]
      operationId: getTimeOffDashboard
      summary: Year calendar days and per-type entitlements for one employee
      parameters:
        - { $ref: '#/components/parameters/EmployeeId' }
        - { $ref: '#/components/parameters/Year' }
      responses:
        '200':
          description: Time off dashboard payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOffDashboardResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pp_at
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    EmployeeId:
      name: employeeId
      in: query
      required: false
      schema: { type: string, format: uuid }
...
    TimeOffTypeId:
      name: timeOffTypeId
      in: query
      required: false
      schema: { type: string, format: uuid }
...
    DateFrom:
      name: dateFrom
      in: query
      required: false
      schema: { type: string, format: date }
    DateTo:
      name: dateTo
      in: query
      required: false
      schema: { type: string, format: date }
...
    Year:
      name: year
      in: query
      required: false
      schema: { type: integer, minimum: 2000, maximum: 2100 }
...
    ActiveOnly:
      name: activeOnly
      in: query
      required: false
      schema: { type: boolean, default: false }
...
    AllocationStatusFilter:
      name: status
      in: query
      required: false
      schema: { $ref: '#/components/schemas/AllocationStatus' }
    RequestStatusFilter:
      name: status
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RequestStatus' }
...
  responses:
    ValidationFailed:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthenticated:
      description: Missing, expired or invalid access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource does not exist or is outside the caller's scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Illegal state transition, overlap or uniqueness violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    UserRole:
      type: string
      enum: [employee, hr_manager, hr_payroll_user, hr_payroll_manager, admin]
...
    TimeOffUnit:
      type: string
      enum: [days, hours]
    TimeOffDurationType:
      type: string
      enum: [full_day, half_day, hours]
    AllocationStatus:
      type: string
      enum: [draft, approved, refused]
    RequestStatus:
      type: string
      enum: [to_approve, approved, refused, cancelled]
...
    Quantity:
      type: string
      description: Decimal day or hour quantity with two fraction digits
      pattern: '^-?\d+\.\d{2}$'
...
    PaginationMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_FAILED, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL]
        message: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }

    # --- shared entities ----------------------------------------------------
    EmployeeRef:
      type: object
      required: [id, firstName, lastName, workEmail]
      properties:
        id: { type: string, format: uuid }
        firstName: { type: string }
        lastName: { type: string }
        workEmail: { type: string, format: email }
        jobPosition: { type: [string, 'null'] }
        departmentName: { type: [string, 'null'] }
...
    WorkingScheduleRef:
      type: object
      required: [id, name, hoursPerWeek]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        hoursPerWeek: { $ref: '#/components/schemas/Quantity' }
...
    TimeOffTypeRef:
      type: object
      required: [id, name, code, unit, color]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        unit: { $ref: '#/components/schemas/TimeOffUnit' }
        color: { type: string }

    # --- auth ---------------------------------------------------------------
...
    TimeOffType:
      type: object
      required: [id, name, code, unit, requiresAllocation, isPaid, approvalRole, color, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        code: { type: string }
        unit: { $ref: '#/components/schemas/TimeOffUnit' }
        requiresAllocation: { type: boolean }
        isPaid: { type: boolean }
        approvalRole: { $ref: '#/components/schemas/UserRole' }
        color: { type: string }
        active: { type: boolean }
    CreateTimeOffTypeRequest:
      type: object
      required: [name, code, unit, color]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        code: { type: string, minLength: 1, maxLength: 8 }
        unit: { $ref: '#/components/schemas/TimeOffUnit' }
        requiresAllocation: { type: boolean, default: true }
        isPaid: { type: boolean, default: true }
        approvalRole: { $ref: '#/components/schemas/UserRole' }
        color: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        active: { type: boolean, default: true }
    UpdateTimeOffTypeRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        unit: { $ref: '#/components/schemas/TimeOffUnit' }
        requiresAllocation: { type: boolean }
        isPaid: { type: boolean }
        approvalRole: { $ref: '#/components/schemas/UserRole' }
        color: { type: string, pattern: '^#[0-9a-fA-F]{6}$' }
        active: { type: boolean }
    TimeOffTypeResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/TimeOffType' }
    TimeOffTypeListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/TimeOffType' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    Allocation:
      type: object
      required: [id, employee, timeOffType, allocated, taken, remaining, validFrom, validTo, status]
      properties:
        id: { type: string, format: uuid }
        employee: { $ref: '#/components/schemas/EmployeeRef' }
        timeOffType: { $ref: '#/components/schemas/TimeOffTypeRef' }
        allocated: { $ref: '#/components/schemas/Quantity' }
        taken: { $ref: '#/components/schemas/Quantity' }
        remaining: { $ref: '#/components/schemas/Quantity' }
        validFrom: { type: string, format: date }
        validTo: { type: string, format: date }
        status: { $ref: '#/components/schemas/AllocationStatus' }
        description: { type: [string, 'null'] }
        approver:
          oneOf:
            - { $ref: '#/components/schemas/EmployeeRef' }
            - { type: 'null' }
    CreateAllocationRequest:
      type: object
      required: [employeeId, timeOffTypeId, allocated, validFrom, validTo]
      properties:
        employeeId: { type: string, format: uuid }
        timeOffTypeId: { type: string, format: uuid }
        allocated: { $ref: '#/components/schemas/Quantity' }
        validFrom: { type: string, format: date }
        validTo: { type: string, format: date }
        description: { type: [string, 'null'], maxLength: 500 }
    UpdateAllocationRequest:
      type: object
      properties:
        allocated: { $ref: '#/components/schemas/Quantity' }
        validFrom: { type: string, format: date }
        validTo: { type: string, format: date }
        description: { type: [string, 'null'], maxLength: 500 }
    AllocationResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Allocation' }
    AllocationListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Allocation' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    TimeOffRequest:
      type: object
      required: [id, employee, timeOffType, startDate, endDate, durationType, durationDays, durationHours, status]
      properties:
        id: { type: string, format: uuid }
        employee: { $ref: '#/components/schemas/EmployeeRef' }
        timeOffType: { $ref: '#/components/schemas/TimeOffTypeRef' }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        durationType: { $ref: '#/components/schemas/TimeOffDurationType' }
        requestedHours: { $ref: '#/components/schemas/Quantity' }
        durationDays: { $ref: '#/components/schemas/Quantity' }
        durationHours: { $ref: '#/components/schemas/Quantity' }
        status: { $ref: '#/components/schemas/RequestStatus' }
        reason: { type: [string, 'null'] }
        refusalReason: { type: [string, 'null'] }
        allocation:
          oneOf:
            - { $ref: '#/components/schemas/Allocation' }
            - { type: 'null' }
        approver:
          oneOf:
            - { $ref: '#/components/schemas/EmployeeRef' }
            - { type: 'null' }
    CreateTimeOffRequestRequest:
      type: object
      required: [employeeId, timeOffTypeId, startDate, endDate, durationType]
      properties:
        employeeId: { type: string, format: uuid }
        timeOffTypeId: { type: string, format: uuid }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        durationType: { $ref: '#/components/schemas/TimeOffDurationType' }
        requestedHours: { $ref: '#/components/schemas/Quantity' }
        reason: { type: [string, 'null'], maxLength: 500 }
    UpdateTimeOffRequestRequest:
      type: object
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        durationType: { $ref: '#/components/schemas/TimeOffDurationType' }
        requestedHours: { $ref: '#/components/schemas/Quantity' }
        reason: { type: [string, 'null'], maxLength: 500 }
        status:
          type: string
          enum: [cancelled]
    RefuseTimeOffRequestRequest:
      type: object
      properties:
        refusalReason: { type: [string, 'null'], maxLength: 500 }
    TimeOffRequestResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/TimeOffRequest' }
    TimeOffRequestListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/TimeOffRequest' }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
    TimeOffCalendarDay:
      type: object
      required: [date, kind]
      properties:
        date: { type: string, format: date }
        kind:
          type: string
          enum: [working, non_working, holiday, leave]
        timeOffTypeId: { type: [string, 'null'], format: uuid }
        color: { type: [string, 'null'] }
        fraction: { $ref: '#/components/schemas/Quantity' }
        label: { type: [string, 'null'] }
        isPending: { type: boolean, default: false }
    TimeOffEntitlement:
      type: object
      required: [timeOffType, allocated, taken, remaining]
      properties:
        timeOffType: { $ref: '#/components/schemas/TimeOffTypeRef' }
        allocated: { $ref: '#/components/schemas/Quantity' }
        taken: { $ref: '#/components/schemas/Quantity' }
        remaining: { $ref: '#/components/schemas/Quantity' }
        pending: { $ref: '#/components/schemas/Quantity' }
    TimeOffDashboard:
      type: object
      required: [employee, year, days, entitlements, unplannedSummary]
      properties:
        employee: { $ref: '#/components/schemas/EmployeeRef' }
        year: { type: integer }
        workingSchedule: { $ref: '#/components/schemas/WorkingScheduleRef' }
        days:
          type: array
          items: { $ref: '#/components/schemas/TimeOffCalendarDay' }
        entitlements:
          type: array
          items: { $ref: '#/components/schemas/TimeOffEntitlement' }
        unplannedSummary: { $ref: '#/components/schemas/UnplannedSummary' }
    UnplannedSummary:
      type: object
      required: [last30Days, last3Months, last6Months, thisYear]
      properties:
        last30Days: { $ref: '#/components/schemas/Quantity' }
        last3Months: { $ref: '#/components/schemas/Quantity' }
        last6Months: { $ref: '#/components/schemas/Quantity' }
        thisYear: { $ref: '#/components/schemas/Quantity' }
    TimeOffDashboardResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/TimeOffDashboard' }

    # --- payroll configuration ----------------------------------------------
...